Refactor menu and userprofile, fix hidden menu title
