--- app/src/components/UserProfile.css
@@ -1,8 +1,0 @@
-ion-row > ion-button {
-  margin-right: 15px;
-  margin-top: 15px;
-}
-
-ion-list > ion-select {
-  padding-left: 20px;
-}
